requirements.txt
gTTS
torchvision
numpy
opencv-python
Flask
